Removed SettingsWindow. Moved SelectionWindow into LeerCopyWPF.Views namespace.
