+++ LeerCopyWPF/Views/SelectionWindow.xaml
@@ -1,4 +1,4 @@
-﻿<Window x:Name="SelectWin" x:Class="LeerCopyWPF.SelectionWindow"
+﻿<Window x:Class="LeerCopyWPF.Views.SelectionWindow"
         xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
         xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
         xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
